move DomainMap to its own file
